refactor:
provider is getting to big, so give it its own submodule and put database
in there.
