+++ src/provider/database.rs
@@ -1,0 +1,298 @@
+use super::{BlobOrCollection, Data};
+use crate::Hash;
+use anyhow::{Context, Result};
+use bytes::Bytes;
+use std::{
+    collections::{BTreeSet, HashMap},
+    fmt, io,
+    path::{Path, PathBuf},
+    result,
+    sync::{Arc, RwLock},
+};
+
+/// Database containing content-addressed data (blobs or collections).
+#[derive(Debug, Clone, Default)]
+pub struct Database(Arc<RwLock<HashMap<Hash, BlobOrCollection>>>);
+
+impl From<HashMap<Hash, BlobOrCollection>> for Database {
+    fn from(map: HashMap<Hash, BlobOrCollection>) -> Self {
+        Self(Arc::new(RwLock::new(map)))
+    }
+}
+
+/// A snapshot of the database.
+///
+/// `E` can be `Infallible` if we take a snapshot from an in memory database,
+/// or `io::Error` if we read a database from disk.
+pub struct Snapshot<E> {
+    /// list of paths we have, hash is the hash of the blob or collection
+    paths: Box<dyn Iterator<Item = (Hash, u64, Option<PathBuf>)>>,
+    /// map of hash to outboard, hash is the hash of the outboard and is unique
+    outboards: Box<dyn Iterator<Item = result::Result<(Hash, Bytes), E>>>,
+    /// map of hash to collection, hash is the hash of the collection and is unique
+    collections: Box<dyn Iterator<Item = result::Result<(Hash, Bytes), E>>>,
+}
+
+impl<E> fmt::Debug for Snapshot<E> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("Snapshot").finish()
+    }
+}
+
+/// An error that can never happen
+#[derive(Debug)]
+pub enum NoError {}
+
+impl From<NoError> for io::Error {
+    fn from(_: NoError) -> Self {
+        unreachable!()
+    }
+}
+
+struct DataPaths {
+    #[allow(dead_code)]
+    data_dir: PathBuf,
+    outboards_dir: PathBuf,
+    collections_dir: PathBuf,
+    paths_file: PathBuf,
+}
+
+impl DataPaths {
+    fn new(data_dir: PathBuf) -> Self {
+        Self {
+            outboards_dir: data_dir.join("collections"),
+            collections_dir: data_dir.join("outboards"),
+            paths_file: data_dir.join("paths.bin"),
+            data_dir: data_dir,
+        }
+    }
+}
+
+/// Using base64 you have all those weird characters like + and /.
+/// So we use hex for file names.
+fn format_hash(hash: &Hash) -> String {
+    hex::encode(hash.as_ref())
+}
+
+/// Parse a hash from a string, e.g. a file name.
+fn parse_hash(hash: &str) -> Result<Hash> {
+    let hash = hex::decode(hash)?;
+    let hash: [u8; 32] = hash.try_into().ok().context("wrong size for hash")?;
+    Ok(Hash::from(hash))
+}
+
+impl Snapshot<io::Error> {
+    /// Load a snapshot from disk.
+    pub fn load(data_dir: impl AsRef<Path>) -> anyhow::Result<Self> {
+        use std::fs;
+        let DataPaths {
+            outboards_dir,
+            collections_dir,
+            paths_file,
+            ..
+        } = DataPaths::new(data_dir.as_ref().to_path_buf());
+        let paths = fs::read(paths_file)?;
+        let paths = postcard::from_bytes::<Vec<(Hash, u64, Option<PathBuf>)>>(&paths)?;
+        let hashes = paths
+            .iter()
+            .map(|(hash, _, _)| *hash)
+            .collect::<BTreeSet<_>>();
+        let outboards = hashes.clone().into_iter().map(move |hash| {
+            let path = outboards_dir.join(format_hash(&hash));
+            fs::read(path).map(|x| (hash, Bytes::from(x)))
+        });
+        let collections = fs::read_dir(collections_dir)?
+            .map(move |entry| {
+                let entry = entry?;
+                let path = entry.path();
+                // skip directories
+                if entry.file_type()?.is_dir() {
+                    tracing::debug!("skipping directory: {:?}", path);
+                    return Ok(None);
+                }
+                // try to get the file name as an OsStr
+                let name = if let Some(name) = path.file_name() {
+                    name
+                } else {
+                    tracing::debug!("skipping unexpected path: {:?}", path);
+                    return Ok(None);
+                };
+                // try to convert into a std str
+                let name = if let Some(name) = name.to_str() {
+                    name
+                } else {
+                    tracing::debug!("skipping unexpected path: {:?}", path);
+                    return Ok(None);
+                };
+                // try to parse the file name as a hash
+                let hash = match parse_hash(name) {
+                    Ok(hash) => hash,
+                    Err(err) => {
+                        tracing::debug!("skipping unexpected path: {:?}: {}", path, err);
+                        return Ok(None);
+                    }
+                };
+                // skip files that are not in the paths file
+                if !hashes.contains(&hash) {
+                    tracing::debug!("skipping unexpected hash: {:?}", hash);
+                    return Ok(None);
+                }
+                // read the collection data and turn it into a Bytes
+                let collection = Bytes::from(fs::read(path)?);
+                io::Result::Ok(Some((hash, collection)))
+            })
+            .filter_map(|x| x.transpose());
+        Ok(Self {
+            paths: Box::new(paths.into_iter()),
+            outboards: Box::new(outboards),
+            collections: Box::new(collections),
+        })
+    }
+}
+
+impl<E> Snapshot<E>
+where
+    io::Error: From<E>,
+{
+    /// Persist the snapshot to disk.
+    pub fn persist(self, data_dir: impl AsRef<Path>) -> io::Result<()> {
+        use std::fs;
+        let DataPaths {
+            outboards_dir,
+            collections_dir,
+            paths_file,
+            ..
+        } = DataPaths::new(data_dir.as_ref().to_path_buf());
+        fs::create_dir_all(&data_dir)?;
+        fs::create_dir_all(&outboards_dir)?;
+        fs::create_dir_all(&collections_dir)?;
+        for item in self.outboards {
+            let (hash, outboard) = item.map_err(Into::into)?;
+            let path = outboards_dir.join(format_hash(&hash));
+            fs::write(path, &outboard)?;
+        }
+        for item in self.collections {
+            let (hash, collection) = item.map_err(Into::into)?;
+            let path = collections_dir.join(format_hash(&hash));
+            fs::write(path, &collection)?;
+        }
+        let mut paths = self.paths.collect::<Vec<_>>();
+        paths.sort_by_key(|(path, _, _)| path.clone());
+        let paths_content = postcard::to_stdvec(&paths).expect("failed to serialize paths file");
+        fs::write(&paths_file, &paths_content)?;
+        Ok(())
+    }
+}
+
+impl Database {
+    /// Load a database from disk.
+    pub fn from_snapshot<E: Into<io::Error>>(snapshot: Snapshot<E>) -> io::Result<Self> {
+        let Snapshot {
+            outboards,
+            collections,
+            paths,
+        } = snapshot;
+        let outboards = outboards
+            .collect::<result::Result<HashMap<_, _>, E>>()
+            .map_err(Into::into)?;
+        let collections = collections
+            .collect::<result::Result<HashMap<_, _>, E>>()
+            .map_err(Into::into)?;
+        let mut db = HashMap::new();
+        for (hash, size, path) in paths {
+            if let Some(path) = path {
+                if let Some(outboard) = outboards.get(&hash) {
+                    db.insert(
+                        hash,
+                        BlobOrCollection::Blob(Data {
+                            outboard: outboard.clone(),
+                            path,
+                            size,
+                        }),
+                    );
+                }
+            }
+        }
+        for (hash, data) in collections {
+            if let Some(outboard) = outboards.get(&hash) {
+                db.insert(
+                    hash,
+                    BlobOrCollection::Collection {
+                        outboard: outboard.clone(),
+                        data,
+                    },
+                );
+            }
+        }
+
+        Ok(Self(Arc::new(RwLock::new(db))))
+    }
+
+    /// take a snapshot of the database
+    pub fn snapshot(&self) -> Snapshot<NoError> {
+        let this = self.0.read().unwrap();
+        let outboards = this
+            .iter()
+            .map(|(k, v)| match v {
+                BlobOrCollection::Blob(Data { outboard, .. }) => (*k, outboard.clone()),
+                BlobOrCollection::Collection { outboard, .. } => (*k, outboard.clone()),
+            })
+            .collect::<Vec<_>>();
+
+        let collections = this
+            .iter()
+            .filter_map(|(k, v)| match v {
+                BlobOrCollection::Blob(_) => None,
+                BlobOrCollection::Collection { data, .. } => Some((*k, data.clone())),
+            })
+            .collect::<Vec<_>>();
+
+        let paths = this
+            .iter()
+            .map(|(k, v)| match v {
+                BlobOrCollection::Blob(Data { path, size, .. }) => (*k, *size, Some(path.clone())),
+                BlobOrCollection::Collection { data, .. } => (*k, data.len() as u64, None),
+            })
+            .collect::<Vec<_>>();
+
+        Snapshot {
+            outboards: Box::new(outboards.into_iter().map(Ok)),
+            collections: Box::new(collections.into_iter().map(Ok)),
+            paths: Box::new(paths.into_iter()),
+        }
+    }
+
+    pub(crate) fn get(&self, key: &Hash) -> Option<BlobOrCollection> {
+        self.0.read().unwrap().get(key).cloned()
+    }
+
+    pub(crate) fn union_with(&self, db: HashMap<Hash, BlobOrCollection>) {
+        let mut inner = self.0.write().unwrap();
+        for (k, v) in db {
+            inner.entry(k).or_insert(v);
+        }
+    }
+
+    /// Iterate over all blobs in the database.
+    pub fn blobs(&self) -> impl Iterator<Item = (Hash, PathBuf, u64)> + 'static {
+        let items = self
+            .0
+            .read()
+            .unwrap()
+            .iter()
+            .filter_map(|(k, v)| match v {
+                BlobOrCollection::Blob(data) => Some((k, data)),
+                BlobOrCollection::Collection { .. } => None,
+            })
+            .map(|(k, data)| (*k, data.path.clone(), data.size))
+            .collect::<Vec<_>>();
+        // todo: make this a proper lazy iterator at some point
+        // e.g. by using an immutable map or a real database that supports snapshots.
+        items.into_iter()
+    }
+
+    #[cfg(test)]
+    pub(crate) fn to_inner(&self) -> HashMap<Hash, BlobOrCollection> {
+        self.0.read().unwrap().clone()
+    }
+}
